(PUP-3632) Update and unify default puppet.conf settings
This commit updates the existing puppet.conf files for various
Linux platforms to point to a single symlink with the new
defaults for vardir, rundir and logdir.

Even though these settings are now sanely set in Puppet itself,
we still need to set these in puppet.conf due to the fact that
the master process runs as the puppet user, so its default
paths are derivatives of ~/.puppet/var. We need to instead use
the new AIO paths to function correctly.